Added login support, began migration support
Login support; salted md5 with session checking. Unsure of session
security in laravel.
Tried to make migration rules, but migration apparently doesn't play
nice with sqlite, so we'll push that closer to production. Favorites
refer to logged in user, and won't show if not logged in. Context aware
navbar, and permission-denying comparison search.